[ETOP] A draft of an Employee Token Option Plan
See README for more details. Not sure if this will be merge-able
as is, but should at least be enough to start a discussion on
if this fits into the current SAFT framework.

Agreements for employees are arguably as important as those for
investors, as employees/founders need to be properly protected,
and accountable to the longevity of these projects.